feat: reorganize settings screens
- Merge Appearance and Interface into a single "Appearance and controls" screen with Theme, Layout, Display, and Behavior groups
- Collect the now playing screen settings into a "Now playing" group under Player and audio, and show the advanced group without folding
- Add a top-level Privacy screen holding listen/search history and proxy settings, and remove the duplicated remote listen history toggle from the sync screen
- Show Lyrics at the top level and display its advanced group without folding
- Replace the content links and the advanced fold in Library and content with Localization and Display groups
- Hide the sync group and "Use your login for browsing" when logged out
- Keep all preference keys unchanged; only item locations change
