+++ app/src/main/java/com/dd3boh/outertune/ui/screens/settings/LibrarySettings.kt
@@ -7,8 +7,6 @@
  */
 package com.dd3boh.outertune.ui.screens.settings
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -20,9 +18,6 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.rounded.ArrowBack
 import androidx.compose.material.icons.automirrored.rounded.PlaylistPlay
 import androidx.compose.material.icons.rounded.FolderCopy
-import androidx.compose.material.icons.rounded.Lyrics
-import androidx.compose.material.icons.rounded.SdCard
-import androidx.compose.material.icons.rounded.Storage
 import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -36,26 +31,15 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.dd3boh.outertune.R
 import com.dd3boh.outertune.constants.FlatSubfoldersKey
-import com.dd3boh.outertune.constants.ProxyEnabledKey
-import com.dd3boh.outertune.constants.ProxyTypeKey
-import com.dd3boh.outertune.constants.ProxyUrlKey
 import com.dd3boh.outertune.constants.ShowLikedAndDownloadedPlaylist
 import com.dd3boh.outertune.constants.TopBarInsets
 import com.dd3boh.outertune.ui.component.ColumnWithContentPadding
-import com.dd3boh.outertune.ui.component.EditTextPreference
-import com.dd3boh.outertune.ui.component.ListPreference
-import com.dd3boh.outertune.ui.component.PreferenceEntry
 import com.dd3boh.outertune.ui.component.PreferenceGroupTitle
-import com.dd3boh.outertune.ui.component.SettingsClickToReveal
 import com.dd3boh.outertune.ui.component.SwitchPreference
 import com.dd3boh.outertune.ui.component.button.IconButton
-import com.dd3boh.outertune.ui.screens.settings.fragments.ListenHistoryFrag
 import com.dd3boh.outertune.ui.screens.settings.fragments.LocalizationFrag
-import com.dd3boh.outertune.ui.screens.settings.fragments.SearchHistoryFrag
 import com.dd3boh.outertune.ui.utils.backToMain
-import com.dd3boh.outertune.utils.rememberEnumPreference
 import com.dd3boh.outertune.utils.rememberPreference
-import java.net.Proxy
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -69,42 +53,12 @@ fun LibrarySettings(
     )
     val (flatSubfolders, onFlatSubfoldersChange) = rememberPreference(FlatSubfoldersKey, defaultValue = true)
 
-    val (proxyEnabled, onProxyEnabledChange) = rememberPreference(key = ProxyEnabledKey, defaultValue = false)
-    val (proxyType, onProxyTypeChange) = rememberEnumPreference(key = ProxyTypeKey, defaultValue = Proxy.Type.HTTP)
-    val (proxyUrl, onProxyUrlChange) = rememberPreference(key = ProxyUrlKey, defaultValue = "host:port")
-
-
     ColumnWithContentPadding(
         modifier = Modifier.fillMaxHeight(),
         columnModifier = Modifier
             .padding(horizontal = 16.dp)
             .verticalScroll(rememberScrollState())
     ) {
-        PreferenceGroupTitle(
-            title = stringResource(R.string.content)
-        )
-
-        ElevatedCard(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            PreferenceEntry(
-                title = { Text(stringResource(R.string.local_player_settings_title)) },
-                icon = { Icon(Icons.Rounded.SdCard, null) },
-                onClick = { navController.navigate("settings/local") }
-            )
-            PreferenceEntry(
-                title = { Text(stringResource(R.string.lyrics_settings_title)) },
-                icon = { Icon(Icons.Rounded.Lyrics, null) },
-                onClick = { navController.navigate("settings/library/lyrics") }
-            )
-            PreferenceEntry(
-                title = { Text(stringResource(R.string.storage)) },
-                icon = { Icon(Icons.Rounded.Storage, null) },
-                onClick = { navController.navigate("settings/storage") }
-            )
-        }
-        Spacer(modifier = Modifier.height(16.dp))
-
         PreferenceGroupTitle(
             title = stringResource(R.string.grp_localization)
         )
@@ -116,74 +70,24 @@ fun LibrarySettings(
         Spacer(modifier = Modifier.height(16.dp))
 
         PreferenceGroupTitle(
-            title = stringResource(R.string.privacy)
+            title = stringResource(R.string.grp_display)
         )
-
         ElevatedCard(
             modifier = Modifier.fillMaxWidth()
         ) {
-            ListenHistoryFrag()
-        }
-        Spacer(modifier = Modifier.height(16.dp))
-
-        ElevatedCard(
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            SearchHistoryFrag()
-        }
-
-        SettingsClickToReveal(stringResource(R.string.advanced)) {
-            ElevatedCard(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                SwitchPreference(
-                    title = { Text(stringResource(R.string.show_liked_and_downloaded_playlist)) },
-                    icon = { Icon(Icons.AutoMirrored.Rounded.PlaylistPlay, null) },
-                    checked = showLikedAndDownloadedPlaylist,
-                    onCheckedChange = onShowLikedAndDownloadedPlaylistChange
-                )
-            }
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ElevatedCard(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                SwitchPreference(
-                    title = { Text(stringResource(R.string.flat_subfolders_title)) },
-                    description = stringResource(R.string.flat_subfolders_description),
-                    icon = { Icon(Icons.Rounded.FolderCopy, null) },
-                    checked = flatSubfolders,
-                    onCheckedChange = onFlatSubfoldersChange
-                )
-            }
-            Spacer(modifier = Modifier.height(16.dp))
-
-            ElevatedCard(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                SwitchPreference(
-                    title = { Text(stringResource(R.string.enable_proxy)) },
-                    checked = proxyEnabled,
-                    onCheckedChange = onProxyEnabledChange
-                )
-
-                AnimatedVisibility(proxyEnabled) {
-                    Column {
-                        ListPreference(
-                            title = { Text(stringResource(R.string.proxy_type)) },
-                            selectedValue = proxyType,
-                            values = listOf(Proxy.Type.HTTP, Proxy.Type.SOCKS),
-                            valueText = { it.name },
-                            onValueSelected = onProxyTypeChange
-                        )
-                        EditTextPreference(
-                            title = { Text(stringResource(R.string.proxy_url)) },
-                            value = proxyUrl,
-                            onValueChange = onProxyUrlChange
-                        )
-                    }
-                }
-            }
+            SwitchPreference(
+                title = { Text(stringResource(R.string.show_liked_and_downloaded_playlist)) },
+                icon = { Icon(Icons.AutoMirrored.Rounded.PlaylistPlay, null) },
+                checked = showLikedAndDownloadedPlaylist,
+                onCheckedChange = onShowLikedAndDownloadedPlaylistChange
+            )
+            SwitchPreference(
+                title = { Text(stringResource(R.string.flat_subfolders_title)) },
+                description = stringResource(R.string.flat_subfolders_description),
+                icon = { Icon(Icons.Rounded.FolderCopy, null) },
+                checked = flatSubfolders,
+                onCheckedChange = onFlatSubfoldersChange
+            )
         }
         Spacer(Modifier.height(96.dp))
     }
