+++ app/src/main/java/com/dd3boh/outertune/ui/screens/settings/LyricsSettings.kt
@@ -33,7 +33,6 @@ import com.dd3boh.outertune.R
 import com.dd3boh.outertune.constants.TopBarInsets
 import com.dd3boh.outertune.ui.component.ColumnWithContentPadding
 import com.dd3boh.outertune.ui.component.PreferenceGroupTitle
-import com.dd3boh.outertune.ui.component.SettingsClickToReveal
 import com.dd3boh.outertune.ui.component.button.IconButton
 import com.dd3boh.outertune.ui.screens.settings.fragments.LyricAdvancedFrag
 import com.dd3boh.outertune.ui.screens.settings.fragments.LyricFormatFrag
@@ -93,9 +92,15 @@ fun LyricsSettings(
         }
         Spacer(modifier = Modifier.height(16.dp))
 
-        SettingsClickToReveal(stringResource(R.string.prefs_advanced)) {
+        PreferenceGroupTitle(
+            title = stringResource(R.string.advanced)
+        )
+        ElevatedCard(
+            modifier = Modifier.fillMaxWidth()
+        ) {
             LyricAdvancedFrag()
         }
+        Spacer(modifier = Modifier.height(16.dp))
     }
