+++ app/src/main/java/com/dd3boh/outertune/ui/screens/settings/fragments/PlayerFrag.kt
@@ -1,15 +1,20 @@
 package com.dd3boh.outertune.ui.screens.settings.fragments
 
+import android.os.Build
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.rounded.QueueMusic
 import androidx.compose.material.icons.automirrored.rounded.VolumeDown
 import androidx.compose.material.icons.automirrored.rounded.VolumeUp
 import androidx.compose.material.icons.rounded.Autorenew
 import androidx.compose.material.icons.rounded.Bedtime
+import androidx.compose.material.icons.rounded.BlurOn
 import androidx.compose.material.icons.rounded.ClearAll
 import androidx.compose.material.icons.rounded.FastForward
 import androidx.compose.material.icons.rounded.GraphicEq
 import androidx.compose.material.icons.rounded.Headset
+import androidx.compose.material.icons.rounded.Lyrics
 import androidx.compose.material.icons.rounded.SkipNext
+import androidx.compose.material.icons.rounded.Swipe
 import androidx.compose.material.icons.rounded.Sync
 import androidx.compose.material.icons.rounded.Timelapse
 import androidx.compose.material.icons.rounded.Visibility
@@ -29,10 +34,19 @@ import com.dd3boh.outertune.constants.AudioNormalizationKey
 import com.dd3boh.outertune.constants.AudioQuality
 import com.dd3boh.outertune.constants.AudioQualityKey
 import com.dd3boh.outertune.constants.AutoLoadMoreKey
+import com.dd3boh.outertune.constants.DEFAULT_PLAYER_BACKGROUND
+import com.dd3boh.outertune.constants.DEFAULT_SHOW_LYRICS_ON_CLICK
+import com.dd3boh.outertune.constants.DEFAULT_SLIDER_STYLE
+import com.dd3boh.outertune.constants.DEFAULT_SWIPE_TO_SKIP
 import com.dd3boh.outertune.constants.IgnoreAudioFocusKey
 import com.dd3boh.outertune.constants.KeepAliveKey
+import com.dd3boh.outertune.constants.PersistentQueueKey
+import com.dd3boh.outertune.constants.PlayerBackgroundStyle
+import com.dd3boh.outertune.constants.PlayerBackgroundStyleKey
 import com.dd3boh.outertune.constants.SeekIncrement
 import com.dd3boh.outertune.constants.SeekIncrementKey
+import com.dd3boh.outertune.constants.ShowLyricsOnClickKey
+import com.dd3boh.outertune.constants.ShowQueueTitleKey
 import com.dd3boh.outertune.constants.SkipOnErrorKey
 import com.dd3boh.outertune.constants.SkipSilenceKey
 import com.dd3boh.outertune.constants.SleepTimerDefaultMinutesKey
@@ -40,7 +54,10 @@ import com.dd3boh.outertune.constants.SleepTimerDefaults
 import com.dd3boh.outertune.constants.SleepTimerFadeDurationKey
 import com.dd3boh.outertune.constants.SleepTimerFadeKey
 import com.dd3boh.outertune.constants.SleepTimerShowOnPlayerKey
+import com.dd3boh.outertune.constants.SliderStyle
+import com.dd3boh.outertune.constants.SliderStyleKey
 import com.dd3boh.outertune.constants.StopMusicOnTaskClearKey
+import com.dd3boh.outertune.constants.SwipeToSkipKey
 import com.dd3boh.outertune.constants.minPlaybackDurKey
 import com.dd3boh.outertune.ui.component.EnumListPreference
 import com.dd3boh.outertune.ui.component.PreferenceEntry
@@ -54,6 +71,7 @@ import androidx.compose.ui.tooling.preview.Preview
 @Composable
 fun PlayerGeneralFrag() {
     val (autoLoadMore, onAutoLoadMoreChange) = rememberPreference(AutoLoadMoreKey, defaultValue = true)
+    val (skipSilence, onSkipSilenceChange) = rememberPreference(key = SkipSilenceKey, defaultValue = false)
 
     val context = LocalContext.current
     val (seekIncrement, onSeekIncrementChange) = rememberEnumPreference(
@@ -77,6 +95,12 @@ fun PlayerGeneralFrag() {
             seekIncrement -> SeekIncrement.getString(context, seekIncrement)
         }
     )
+    SwitchPreference(
+        title = { Text(stringResource(R.string.skip_silence)) },
+        icon = { Icon(painterResource(R.drawable.skip_next), null) },
+        checked = skipSilence,
+        onCheckedChange = onSkipSilenceChange
+    )
 }
 
 @Composable
@@ -89,6 +113,10 @@ fun AudioQualityFrag() {
     val (audioQuality, onAudioQualityChange) = rememberEnumPreference(
         key = AudioQualityKey,
         defaultValue = AudioQuality.AUTO
+    )
+    val (audioNormalization, onAudioNormalizationChange) = rememberPreference(
+        key = AudioNormalizationKey,
+        defaultValue = true
     )
 
     EnumListPreference(
@@ -104,36 +132,84 @@ fun AudioQualityFrag() {
             }
         }
     )
-
-}
-
-@Composable
-fun AudioEffectsFrag() {
-    val (skipSilence, onSkipSilenceChange) = rememberPreference(key = SkipSilenceKey, defaultValue = false)
-
-    val (audioNormalization, onAudioNormalizationChange) = rememberPreference(
-        key = AudioNormalizationKey,
-        defaultValue = true
-    )
-
     SwitchPreference(
         title = { Text(stringResource(R.string.audio_normalization)) },
         icon = { Icon(Icons.AutoMirrored.Rounded.VolumeUp, null) },
         checked = audioNormalization,
         onCheckedChange = onAudioNormalizationChange
     )
-    SwitchPreference(
-        title = { Text(stringResource(R.string.skip_silence)) },
-        icon = { Icon(painterResource(R.drawable.skip_next), null) },
-        checked = skipSilence,
-        onCheckedChange = onSkipSilenceChange
+}
+
+@Composable
+fun NowPlayingFrag() {
+    val (playerBackground, onPlayerBackgroundChange) = rememberEnumPreference(
+        key = PlayerBackgroundStyleKey,
+        defaultValue = DEFAULT_PLAYER_BACKGROUND
+    )
+    val availableBackgroundStyles = PlayerBackgroundStyle.entries.filter {
+        it != PlayerBackgroundStyle.BLUR || Build.VERSION.SDK_INT >= Build.VERSION_CODES.S
+    }
+    val (showQueueTitle, onShowQueueTitleChange) = rememberPreference(ShowQueueTitleKey, defaultValue = true)
+    val (sliderStyle, onSliderStyleChange) = rememberEnumPreference(SliderStyleKey, defaultValue = DEFAULT_SLIDER_STYLE)
+    val (swipeToSkip, onSwipeToSkipChange) = rememberPreference(SwipeToSkipKey, defaultValue = DEFAULT_SWIPE_TO_SKIP)
+    val (showLyricsOnClick, onShowLyricsOnClickChange) = rememberPreference(
+        ShowLyricsOnClickKey,
+        defaultValue = DEFAULT_SHOW_LYRICS_ON_CLICK
     )
 
+    EnumListPreference(
+        title = { Text(stringResource(R.string.player_background_style)) },
+        icon = { Icon(Icons.Rounded.BlurOn, null) },
+        selectedValue = playerBackground,
+        onValueSelected = onPlayerBackgroundChange,
+        valueText = {
+            when (it) {
+                PlayerBackgroundStyle.FOLLOW_THEME -> stringResource(R.string.player_background_default)
+                PlayerBackgroundStyle.GRADIENT -> stringResource(R.string.player_background_gradient)
+                PlayerBackgroundStyle.BLUR -> stringResource(R.string.player_background_blur)
+            }
+        },
+        values = availableBackgroundStyles
+    )
+    EnumListPreference(
+        title = { Text(stringResource(R.string.slider_style_title)) },
+        icon = { Icon(Icons.Rounded.GraphicEq, null) },
+        selectedValue = sliderStyle,
+        onValueSelected = onSliderStyleChange,
+        valueText = {
+            when (it) {
+                SliderStyle.SQUIGGLY -> stringResource(R.string.slider_style_squiggly)
+                SliderStyle.DEFAULT -> stringResource(R.string.slider_style_default)
+                SliderStyle.SLIM -> stringResource(R.string.slider_style_slim)
+            }
+        }
+    )
+    SwitchPreference(
+        title = { Text(stringResource(R.string.swipe_to_skip_title)) },
+        description = stringResource(R.string.swipe_to_skip_description),
+        icon = { Icon(Icons.Rounded.Swipe, null) },
+        checked = swipeToSkip,
+        onCheckedChange = onSwipeToSkipChange
+    )
+    SwitchPreference(
+        title = { Text(stringResource(R.string.show_queue_title)) },
+        icon = { Icon(Icons.AutoMirrored.Rounded.QueueMusic, null) },
+        checked = showQueueTitle,
+        onCheckedChange = onShowQueueTitleChange
+    )
+    SwitchPreference(
+        title = { Text(stringResource(R.string.tap_artwork_to_show_lyrics_title)) },
+        description = stringResource(R.string.tap_artwork_to_show_lyrics_description),
+        icon = { Icon(Icons.Rounded.Lyrics, null) },
+        checked = showLyricsOnClick,
+        onCheckedChange = onShowLyricsOnClickChange
+    )
 }
 
 @Composable
 fun PlaybackBehaviourFrag() {
     val keepAlive by rememberPreference(key = KeepAliveKey, defaultValue = false)
+    val (persistentQueue, onPersistentQueueChange) = rememberPreference(key = PersistentQueueKey, defaultValue = true)
     val (minPlaybackDur, onMinPlaybackDurChange) = rememberPreference(minPlaybackDurKey, defaultValue = 30)
     val (skipOnErrorKey, onSkipOnErrorChange) = rememberPreference(key = SkipOnErrorKey, defaultValue = false)
     val (stopMusicOnTaskClear, onStopMusicOnTaskClearChange) = rememberPreference(
@@ -171,6 +247,13 @@ fun PlaybackBehaviourFrag() {
         icon = { Icon(Icons.Rounded.Headset, null) },
         checked = ignoreAudioFocus,
         onCheckedChange = onIgnoreAudioFocusChange
+    )
+    SwitchPreference(
+        title = { Text(stringResource(R.string.persistent_queue)) },
+        description = stringResource(R.string.persistent_queue_desc_ot),
+        icon = { Icon(Icons.AutoMirrored.Rounded.QueueMusic, null) },
+        checked = persistentQueue,
+        onCheckedChange = onPersistentQueueChange
     )
 
     /**
@@ -291,8 +374,8 @@ private fun AudioQualityFragPreview() {
 
 @Preview(showBackground = true)
 @Composable
-private fun AudioEffectsFragPreview() {
-    AudioEffectsFrag()
+private fun NowPlayingFragPreview() {
+    NowPlayingFrag()
 }
 
 @Preview(showBackground = true)
