+++ app/src/main/res/values/strings-ot.xml
@@ -80,8 +80,10 @@ while also allowing users to translate our app as well. New, specific strings to
     <string name="settings_debug">Debug</string>
 
     <string name="grp_account_sync">Account and sync</string>
+    <string name="grp_appearance_controls">Appearance and controls</string>
     <string name="grp_audio">Audio</string>
     <string name="grp_behavior">Behavior</string>
+    <string name="grp_display">Display</string>
     <string name="grp_extra_scanner_settings">Additional scanner settings</string>
     <string name="grp_general">General</string>
     <string name="grp_interface">Interface</string>
@@ -93,6 +95,7 @@ while also allowing users to translate our app as well. New, specific strings to
     <string name="grp_lyrics_source">Lyric sources</string>
     <string name="grp_manual_scanner">Manual scanner</string>
     <string name="grp_more_settings">More settings</string>
+    <string name="grp_now_playing">Now playing</string>
     <string name="grp_proxy">Proxy</string>
     <string name="grp_sync">Sync</string>
